TG-500: Moved example generation InputFileInput
No more forward ref.
